extract `\QIT_CLI_Tests\ZipBuilderHelper`
